tests/wpt/tests/css/css-conditional/container-queries/no-layout-containment-scroll.html

<!DOCTYPE html>
<title>Scrollable overflow should propagate from a container-type subtree.</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/10544#issuecomment-2248438355">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('#test')">
  <div id=log></div>
  <div id="test" data-expected-scroll-height="200">
    <div style="container-type: inline-size; height: 100px;">
      <div style="height: 200px;"></div>
    </div>
  </div>
</body>
