LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/no-layout-containment-scroll-expected.txt


FAIL #test 1 assert_equals:
<div id="test" data-expected-scroll-height="200">
    <div style="container-type: inline-size; height: 100px;">
      <div style="height: 200px;"></div>
    </div>
  </div>
scrollHeight expected 200 but got 100
